@@ -16,12 +16,12 @@
 
 // React core.
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // Firebase.
 import firebase from 'firebase/compat/app';
 import 'firebase/compat/auth';
-import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
+import { FirebaseAuth as StyledFirebaseAuth } from 'react-firebaseui';
 
 // Styles
 import styles from './app.css'; // This uses CSS modules.
@@ -96,4 +96,4 @@ class App extends React.Component {
 }
 
 // Load the app in the browser.
-ReactDOM.render(<App/>, document.getElementById('app'));
+createRoot(document.getElementById('app')).render(<App />);
